+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+*  funcionalidades basicas de la api fake store (productos,categorias,carrito de compras,login,crear usuario,actualizar carrito y usuario)